finished ui and changed some tests

+++ src/main/model/Bike.java
@@ -2,13 +2,13 @@ package model;
 
 public class Bike extends Workout {
 
-    private int avgBikeSpeed;            // average speed in kilometers per hour (integer values only)
+    private double avgBikeSpeed;            // average speed in kilometers per hour (integer values only)
 
 
     // REQUIRES: bikeName is not an empty string, todaysDate is not an empty string, bikeHR > 0, bikeTime > 0
     //           bikeSpeed > 0, bikePerceivedDifficulty is between 1 and 10, bikeDistance > 0
     // EFFECTS: creates new bike object with given parameters
-    public Bike(String bikeName, String todaysDate, int bikeHR, int bikeTime, int bikeSpeed,
+    public Bike(String bikeName, String todaysDate, int bikeHR, int bikeTime, double bikeSpeed,
                 int bikePerceivedDifficulty, double bikeDistance) {
 
         super(bikeName, todaysDate, bikeHR, bikeTime, bikePerceivedDifficulty, bikeDistance);
@@ -23,7 +23,7 @@ public class Bike extends Workout {
     }
 
     // EFFECTS: returns average bike speed
-    public int getAvgBikeSpeed() {
+    public double getAvgBikeSpeed() {
         return this.avgBikeSpeed;
     }
